+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Correção do problema de não exibr botões de compras da etiqueta na listagem de pedidos
 - Correção no processo de finalização de compras de produtos virtuais
 - Ajuste para usar valores com centavos nas taxas extras de envios
+- Ajuste no aviso que o usuário não selecionou um método do Melhor Envio mesmo selecioando um método do Melhor Envio
 
 ### [2.9.7]
 - Ajuste na finalização de pedidos de produtos virtuais e remoção de pedidos virtuais na listagem de pedidos do plugin.